Users & Admin

+++ app/Http/Controllers/CommentController.php
@@ -2,8 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
+use Gate;
 use Illuminate\Http\Request;
 use App\Models\Comment;
+
 
 class CommentController extends Controller
 {
@@ -17,7 +20,7 @@ class CommentController extends Controller
         $comment->name = request('name');
         $comment->desc = request('desc');
         $comment->article_id = request('article_id');
-        $comment->user_id = 1;
+        $comment->user_id = Auth::id();
         $comment->save();
         return redirect()->back();
     }
@@ -25,11 +28,13 @@ class CommentController extends Controller
     public function edit($id)
     {
         $comment = Comment::findOrFail($id);
+        Gate::authorize('update-comment', ['comment' => $comment]);
         return view('comments.update', ['comment' => $comment]);
     }
 
     public function update(Request $request, Comment $comment)
     {
+        Gate::authorize('update-comment', ['comment' => $comment]);
         $request->validate([
             'name' => 'required|min:3',
             'desc' => 'required|max:256'
@@ -43,6 +48,7 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::findOrFail($id);
+        Gate::authorize('update-comment', ['comment' => $comment]);
         $comment->delete();
         return redirect()->route('articles.show', ['article' => $comment->article_id])->with('status', 'Delete success');
     }
